client: always add one new capsnap if Fb is used and Fw is not used
If we set the 'writing' to 1 when the 'Fb' caps is used then later
if we have any dirty caps it will be skipped and will reuse the
existing capsnap, which is incorrect.

At the same time trigger to flush the buffer when making snapshot
and if the Fb is being used.

Fixes: https://tracker.ceph.com/issues/59343
Signed-off-by: Xiubo Li <xiubli@redhat.com>