mgr/cephadm: update package to use ArgumentSpec based types
Update the cephadm mgr module to match the changes to use
ArgumentSpec based types for extra_container_args and extra_entrypoint
args.

Update orchestrator/_interface.py to match the changes to use
ArgumentSpec based types for extra_container_args and extra_entrypoint args.

The job of splitting up arguments is no longer done here but
lies in a method of the ArgumentSpec type.

Signed-off-by: John Mulligan <jmulligan@redhat.com>