qa: fix the threshold value
CInode::maybe_ephemeral_rand() generates a random number,
which should be <= minimum of mds_export_ephemeral_random_max and
ceph.dir.pin.random xattr (threshold).
So we should have a higher threshold for the balancer/random test,
so that the counters could make an increment 80% of the time.
Reducing the value reduces the chances of hitting
l_mds_imported/l_mds_exported.

Fixes: https://tracker.ceph.com/issues/72386
Signed-off-by: Jos Collin <jcollin@redhat.com>